requirements.txt
# Core Framework
fastapi
uvicorn
pydantic
pydantic-settings

# Database
sqlalchemy
alembic
psycopg2-binary
asyncpg

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart

# SMS Authentication
twilio

# Utilities
python-dotenv
httpx
email-validator
